+++ pkg/analysis/stdlib/testdata/abs/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = abs(-5)
+}