+++ pkg/analysis/stdlib/testdata/ceil/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = ceil(3.2)
+}