+++ pkg/analysis/stdlib/testdata/chunklist/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = chunklist(["a", "b", "c", "d", "e"], 2)
+}