+++ pkg/analysis/stdlib/testdata/join/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = join("-", ["a", "b", "c"])
+}