+++ pkg/analysis/stdlib/testdata/pow/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = pow(2, 10)
+}