+++ pkg/analysis/stdlib/testdata/range/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = range(1, 4)
+}