+++ pkg/analysis/stdlib/testdata/replace_literal/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = replace("foo bar foo", "foo", "baz")
+}