+++ pkg/analysis/stdlib/testdata/reverse/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = reverse(["a", "b", "c"])
+}