+++ pkg/analysis/stdlib/testdata/slice/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = slice(["a", "b", "c", "d", "e"], 1, 4)
+}