+++ pkg/analysis/stdlib/testdata/sort/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = sort(["c", "a", "b"])
+}