+++ pkg/analysis/stdlib/testdata/split/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = split(",", "a,b,c")
+}