+++ pkg/analysis/stdlib/testdata/substr/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = substr("abcdefgh", 2, 3)
+}