+++ pkg/analysis/stdlib/testdata/trim/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = trim("!?hello?!", "!?")
+}