+++ pkg/analysis/stdlib/testdata/trimprefix/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = trimprefix("hello-world", "hello-")
+}