+++ pkg/analysis/stdlib/testdata/trimspace/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = trimspace("  hello  ")
+}