+++ pkg/analysis/stdlib/testdata/trimsuffix/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = trimsuffix("hello-world", "-world")
+}