+++ pkg/diff/testdata/tracked_eval_format_collapses/old.tf
@@ -1,0 +1,3 @@
+locals {
+  image = "ec2-small-v3" # tflens:track: AMI image identifier
+}